Adding websocket support
